// src/interfaces/IStore.ts
export interface IStore {
    firstName: string;
    lastName: string;
    email: string;
    industry: string;
    storeType: string;
    storeName: string;
    sortURL: string;
    status: string;
    description: string;
    address1: string;
    address2: string;
    phoneNumber: string;
    state: string;
    logo: string;
    country: string;
    postalCodes: string[];
    date: Date;
}